Refactored to use django-configurations fully, removed python-decouple
