+++ .env.template
@@ -1,4 +1,4 @@
-SECRET_KEY=...
-DEBUG=false
-PUBLIC_URL=https://fiesta.localhost
+DJANGO_DEBUG=false
 DJANGO_LOG_LEVEL=INFO
+DJANGO_SECRET_KEY=
+PUBLIC_URL=https://fiesta.localhost
