+++ docker-compose.yml
@@ -10,7 +10,6 @@ services:
       - 8000
     env_file:
       - ./fiesta/.env.base
-    secrets:
       - .env
     build:
       context: .
@@ -131,10 +130,6 @@ services:
     env_file:
       - .env
       - wiki/.env.base
-
-secrets:
-  .env:
-    file: ./.env
 
 volumes:
   postgres_data:
