+++ fiesta/.env.base
@@ -1,3 +1,3 @@
-BUILD_DIR=/usr/src/build
-STATIC_DIR=/usr/src/static
-MEDIA_DIR=/usr/src/media
+DJANGO_BUILD_DIR=/usr/src/build
+DJANGO_STATIC_ROOT=/usr/src/static
+DJANGO_MEDIA_ROOT=/usr/src/media
