+++ fiesta/fiesta/settings/__init__.py
@@ -1,42 +1,16 @@
-"""
-Django settings for fiesta project.
-
-Generated by 'django-admin startproject' using Django 4.0.1.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/4.0/topics/settings/
-
-For the full list of settings and their values, see
-https://docs.djangoproject.com/en/4.0/ref/settings/
-"""
-
-from pathlib import Path
-
 from configurations import Configuration
-from configurations.values import BooleanValue
-from decouple import AutoConfig
-
-# to load .env from folder, where docker stores secrets files
-
-config = AutoConfig("/run/secrets/")
-
-
-class BaseConfigurationProtocol:
-    DEBUG: bool
-    BASE_DIR: Path
-
 
 from .auth import AuthConfigMixin
 from .db import DatabaseConfigMixin
 from .files import FilesConfigMixin
 from .logging import LoggingConfigMixin
+from .project import ProjectConfigMixin
 from .security import SecurityConfigMixin
 from .templates import TemplatesConfigMixin
-from .app import AppConfigMixin
 
 
 class Base(
-    AppConfigMixin,
+    ProjectConfigMixin,
     AuthConfigMixin,
     DatabaseConfigMixin,
     FilesConfigMixin,
@@ -49,16 +23,23 @@ class Base(
 
 
 class Development(Base):
-    DEBUG = BooleanValue(default=True)
+    DEBUG = True
+    DEBUG_PROPAGATE_EXCEPTIONS = False
+
+    INTERNAL_IPS = type("ContainsAll", (), {"__contains__": lambda *_: True})()
 
     def INSTALLED_APPS(self):
         return super().INSTALLED_APPS + ['debug_toolbar']
 
-    INTERNAL_IPS = type("ContainsAll", (), {"__contains__": lambda *_: True})()
-
     def MIDDLEWARE(self):
-        middlewares = super().MIDDLEWARE[:]
+        middlewares = super().MIDDLEWARE
         middlewares.insert(1, "debug_toolbar.middleware.DebugToolbarMiddleware")
         return middlewares
 
-    # DEBUG_PROPAGATE_EXCEPTIONS = False
+
+class LocalProduction(Base):
+    DEBUG = False
+
+
+class Production(Base):
+    DEBUG = False
