+++ fiesta/fiesta/settings/logging.py
@@ -1,32 +1,35 @@
-from . import config
+from configurations.values import Value
 
 
 class LoggingConfigMixin:
-    LOGGING = {
-        "version": 1,
-        "disable_existing_loggers": False,
-        "formatters": {
-            "verbose": {
-                "format": "[%(asctime)s] %(levelname)s [%(name)s:%(lineno)s] %(message)s",
-                "datefmt": "%d/%b/%Y %H:%M:%S",
+    LOG_LEVEL: str = Value('INFO')
+
+    def LOGGING(self):
+        return {
+            "version": 1,
+            "disable_existing_loggers": False,
+            "formatters": {
+                "verbose": {
+                    "format": "[%(asctime)s] %(levelname)s [%(name)s:%(lineno)s] %(message)s",
+                    "datefmt": "%d/%b/%Y %H:%M:%S",
+                },
+                "simple": {"format": "%(levelname)s %(message)s"},
             },
-            "simple": {"format": "%(levelname)s %(message)s"},
-        },
-        "handlers": {
-            "console": {
-                "class": "logging.StreamHandler",
-                "formatter": "verbose",
+            "handlers": {
+                "console": {
+                    "class": "logging.StreamHandler",
+                    "formatter": "verbose",
+                },
             },
-        },
-        "root": {
-            "handlers": ["console"],
-            "level": "INFO",
-        },
-        "loggers": {
-            "django": {
+            "root": {
                 "handlers": ["console"],
-                "level": config("DJANGO_LOG_LEVEL", "INFO"),
-                "propagate": False,
+                "level": "INFO",
             },
-        },
-    }
+            "loggers": {
+                "django": {
+                    "handlers": ["console"],
+                    "level": self.LOG_LEVEL,
+                    "propagate": False,
+                },
+            },
+        }
