+++ fiesta/fiesta/settings/security.py
@@ -1,4 +1,9 @@
+from configurations.values import SecretValue
+
+
 class SecurityConfigMixin:
+    SECRET_KEY = SecretValue()
+
     SECURE_PROXY_SSL_HEADER = "HTTP_X_FORWARDED_SSL", "on"
     CSRF_TRUSTED_ORIGINS = ["https://*.localhost"]
